chore: change bulk skills action button to skills column
