+++ plugin-hrm-form/src/components/teamsView/SkillsColumn.tsx
@@ -15,20 +15,77 @@
  */
 
 import React from 'react';
-import { WorkersDataTable, ColumnDefinition, Template } from '@twilio/flex-ui';
+import { createPortal } from 'react-dom';
+import { useDispatch, useSelector } from 'react-redux';
+import { WorkersDataTable, ColumnDefinition, Template, styled } from '@twilio/flex-ui';
 import { Tooltip } from '@material-ui/core';
+import { ArrowDropDown, ArrowDropUp, KeyboardArrowRight } from '@material-ui/icons';
 
-import { OpaqueText } from '../../styles';
+import { Flex, Column, OpaqueText, Row, Box } from '../../styles';
+import { MultiSelectButton } from '../../styles/filters';
 import { SkillsList, StyledChip } from './styles';
 import { sortSkills } from './teamsViewSorting';
+import { newTeamsViewSelectOperation, TeamsViewState } from '../../states/teamsView';
+import { newOpenModalAction } from '../../states/routing/actions';
+import type { RootState } from '../../states';
+import { namespace, teamsViewBase } from '../../states/storeNamespaces';
+import { standaloneTask } from '../../types/types';
 
 const MAX_SKILL_LENGTH = 12;
+
+const SkillsColumnHeader: React.FC = () => {
+  const dispatch = useDispatch();
+  const { selectedWorkers } = useSelector((state: RootState) => state[namespace][teamsViewBase]);
+
+  const [isOpen, setIsOpen] = React.useState(false);
+  const buttonRef = React.useRef(null);
+
+  const boxRef = React.useRef(null);
+
+  return (
+    <>
+      <Row ref={boxRef}>
+        <Template code="Skills" />
+        <Box marginLeft="10px" style={{ ...(!selectedWorkers.size && { opactity: 0, visibility: 'hidden' }) }}>
+          <MultiSelectButton
+            isOpened={isOpen}
+            isActive={Boolean(selectedWorkers.size)}
+            disabled={!selectedWorkers.size}
+            type="button"
+            onClick={e => {
+              e.stopPropagation();
+              setIsOpen(prev => !prev);
+            }}
+            ref={buttonRef}
+            style={{ textDecorationLine: 'none' }}
+          >
+            <Template code="Actions" />
+            <Flex marginLeft="15px">
+              {isOpen && <ArrowDropUp />}
+              {!isOpen && <ArrowDropDown />}
+            </Flex>
+          </MultiSelectButton>
+        </Box>
+      </Row>
+      <DropdownPortal
+        anchorRef={buttonRef}
+        open={isOpen}
+        onClickAction={operation => e => {
+          e.stopPropagation();
+          dispatch(newTeamsViewSelectOperation(operation));
+          dispatch(newOpenModalAction({ route: 'teams', subroute: 'select-skills' }, standaloneTask.taskSid));
+          setIsOpen(false);
+        }}
+      />
+    </>
+  );
+};
 
 export const setUpSkillsColumn = () => {
   WorkersDataTable.Content.add(
     <ColumnDefinition
       key="skills"
-      header="Skills"
+      header={<SkillsColumnHeader />}
       sortingFn={sortSkills}
       style={{ width: 'auto !important' }}
       content={item => <SkillsCell item={item} />}
@@ -67,5 +124,81 @@ const SkillsCell = ({ item }) => {
         ),
       )}
     </SkillsList>
+  );
+};
+
+const StyledButton = styled.button`
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  width: 100%;
+  padding: 8px 12px;
+  background: white;
+  border: none;
+  text-align: left;
+  cursor: pointer;
+  font-family: sans-serif;
+  transition: background 0.2s ease,
+
+  &:hover {
+    background: #f8f8f8;
+  }
+`;
+const DropDownButton: React.FC<{ onClick: React.MouseEventHandler<HTMLButtonElement> }> = ({ onClick, children }) => {
+  return (
+    <StyledButton type="button" onClick={onClick}>
+      {children}
+    </StyledButton>
+  );
+};
+const DropdownPortal: React.FC<{
+  anchorRef: any;
+  open: boolean;
+  onClickAction: (operation: TeamsViewState['operation']) => React.MouseEventHandler<HTMLButtonElement>;
+}> = ({ anchorRef, open, onClickAction }) => {
+  const [coords, setCoords] = React.useState(null);
+
+  // Update dropdown position when opened or resized
+  React.useLayoutEffect(() => {
+    if (open && anchorRef.current) {
+      const rect = anchorRef.current.getBoundingClientRect();
+
+      setCoords({
+        top: rect.bottom,
+        left: rect.left,
+      });
+    }
+  }, [anchorRef, open]);
+
+  if (!open || !coords) return null;
+
+  return createPortal(
+    <div
+      style={{
+        position: 'fixed',
+        top: coords.top,
+        left: coords.left,
+        marginTop: '5px',
+        background: 'white',
+        boxSizing: 'border-box',
+        width: '230px',
+        border: '1px solid lightgray',
+        zIndex: 9999,
+      }}
+      role="dialog"
+      aria-labelledby="dialog-title"
+    >
+      <Column>
+        <DropDownButton onClick={onClickAction('enable')}>
+          <Template code="Enable Skills" />
+          <KeyboardArrowRight style={{ marginLeft: '20px' }} />
+        </DropDownButton>
+        <DropDownButton onClick={onClickAction('disable')}>
+          <Template code="Disable Skills" />
+          <KeyboardArrowRight style={{ marginLeft: '20px' }} />
+        </DropDownButton>
+      </Column>
+    </div>,
+    document.body,
   );
 };
